fix: account initialize failure when looking up lego dns provider
after updating to lego/v4, I had to rewrite the getDNSChallengeProvider
logic to initialize the configuration. The vault write `acme/account`
hit a code path that used the older logic that didnt properly initialize
the provider configuration before searching for the provider. This
caused errors that prevented token rotation which eventually lead to all
new dns challenges failing because of expired dns token.